feat: add rclone gateway
This is a higher-level concept that in theory allows for deploying
rclone in "S3 gateway" mode to access storage behind it.

The use case is to provide local, limited access (hence the netpols) to
crypt-wrapped backends. The primary and first consumer is Velero backup
repositories.

This is pretty close to working, but does require you to create the
rclone config yourself and put it in a secret.